requirements.txt
streamlit>=1.35.0,<2.0.0
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<3.0.0
sentence-transformers>=2.2.0,<4.0.0
faiss-cpu>=1.7.0,<2.0.0
tiktoken>=0.5.0,<1.0.0
scikit-learn>=1.3.0,<2.0.0
plotly>=5.15.0,<7.0.0
openai>=1.0.0,<2.0.0
